# backend-api.openapi.yaml
openapi: 3.0.3
info:
  title: Car Service Admin API
  version: 2.0.0
  description: |
    Backend contract for the Car Service Admin dashboard.

    The frontend currently covers authentication, onboarding, dashboard widgets,
    bookings, payments, payment exports, reports, storage inventory, business
    settings, billing account verification, notification preferences, inbox,
    feedback, help center content, and location search.

    API JSON uses camelCase. The recommended database schema uses snake_case.
    Dates should be stored and exchanged as ISO-8601 strings. The frontend can
    format those dates for display.
  contact:
    name: Car Service Admin
servers:
  - url: http://localhost:3333
    description: Local backend development
  - url: https://api.car-service-admin.example.com
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /onboarding/owner-information:
    put:
      tags: [Onboarding]
      operationId: saveOwnerInformation
      summary: Save owner information onboarding step
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnerInformationDto"
      responses:
        "200":
          description: Step saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingStepResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
  /onboarding/business-information:
    put:
      tags: [Onboarding]
      operationId: saveBusinessInformation
      summary: Save business information onboarding step
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessInformationDto"
      responses:
        "200":
          description: Step saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingStepResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
  /onboarding/bank-details:
    put:
      tags: [Onboarding]
      operationId: saveOnboardingBankDetails
      summary: Save bank details collected during onboarding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankAccountCreateDto"
      responses:
        "200":
          description: Step saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingStepResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
  /onboarding/two-factor-auth:
    put:
      tags: [Onboarding]
      operationId: saveOnboardingMfa
      summary: Save MFA settings collected during onboarding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MfaUpdateDto"
      responses:
        "200":
          description: Step saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingStepResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
  /onboarding/submit:
    post:
      tags: [Onboarding]
      operationId: submitOnboarding
      summary: Mark onboarding ready for review or activation
      responses:
        "200":
          description: Onboarding submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingProgress"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /vehicles:
    get:
      tags: [Vehicles]
      operationId: listVehicles
      summary: List vehicles for customers in the current business
      parameters:
        - in: query
          name: customerId
          schema:
            type: string
            format: uuid
        - in: query
          name: search
          schema:
            type: string
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Paginated vehicles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedVehicles"
    post:
      tags: [Vehicles]
      operationId: createVehicle
      summary: Create a vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleCreateDto"
      responses:
        "201":
          description: Created vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/{vehicleId}:
    get:
      tags: [Vehicles]
      operationId: getVehicle
      summary: Get vehicle details
      parameters:
        - $ref: "#/components/parameters/VehicleIdPath"
      responses:
        "200":
          description: Vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Vehicles]
      operationId: updateVehicle
      summary: Update vehicle details
      parameters:
        - $ref: "#/components/parameters/VehicleIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleUpdateDto"
      responses:
        "200":
          description: Updated vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
...
  /payments:
    get:
      tags: [Payments]
      operationId: listPayments
      summary: List payment history with filters, search, sort, and pagination
      parameters:
        - in: query
          name: method
          schema:
            $ref: "#/components/schemas/PaymentMethod"
        - in: query
          name: kind
          schema:
            $ref: "#/components/schemas/PaymentKind"
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [id, customer, item, method, date, amount]
        - $ref: "#/components/parameters/SortOrderQuery"
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Paginated payments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPayments"
  /payments/summary:
    get:
      tags: [Payments]
      operationId: getPaymentSummary
      summary: Get payment summary cards
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
      x-cache:
        store: redis
        ttlSeconds: 120
        key: "business:{businessId}:payments:summary:{startDate}:{endDate}"
      responses:
        "200":
          description: Payment summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSummary"
  /payments/{paymentId}:
    get:
      tags: [Payments]
      operationId: getPayment
      summary: Get payment details
      parameters:
        - $ref: "#/components/parameters/PaymentIdPath"
      responses:
        "200":
          description: Payment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          $ref: "#/components/responses/NotFound"
  /payments/{paymentId}/refund:
    post:
      tags: [Payments]
      operationId: refundPayment
      summary: Refund a payment
      parameters:
        - $ref: "#/components/parameters/PaymentIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundPaymentDto"
      responses:
        "201":
          description: Refund created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRefund"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /payments/batch/refund:
    post:
      tags: [Payments]
      operationId: refundPayments
      summary: Refund selected payments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchRefundPaymentsDto"
      responses:
        "200":
          description: Batch refund result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchMutationResult"
  /payments/export-jobs:
    post:
      tags: [Payments]
      operationId: createPaymentExportJob
      summary: Create an async payment export job
      x-queue:
        name: payment-export
        reason: Large Excel, PDF, Word, Google Docs, and dashboard exports should run outside the request cycle.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentExportCreateDto"
      responses:
        "202":
          description: Export job queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportJob"
        "400":
          $ref: "#/components/responses/ValidationError"
  /payments/export-jobs/{jobId}:
    get:
      tags: [Payments]
      operationId: getPaymentExportJob
      summary: Get export job status and download metadata
      parameters:
        - $ref: "#/components/parameters/JobIdPath"
      responses:
        "200":
          description: Export job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportJob"
        "404":
          $ref: "#/components/responses/NotFound"
  /payments/export-jobs/{jobId}/download:
    get:
      tags: [Payments]
      operationId: downloadPaymentExport
      summary: Download a completed payment export file
      parameters:
        - $ref: "#/components/parameters/JobIdPath"
      responses:
        "200":
          description: Export file
          content:
            application/vnd.ms-excel:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
            application/msword:
              schema:
                type: string
                format: binary
            text/html:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Export is not complete yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /support/help-center:
    get:
      tags: [Support]
      operationId: getHelpCenter
      summary: Get help center categories and articles
      security: []
      x-cache:
        store: redis
        ttlSeconds: 3600
        key: "support:help-center:{locale}"
      parameters:
        - in: query
          name: locale
          schema:
            type: string
            enum: [en, ar]
            default: en
      responses:
        "200":
          description: Help center content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelpCenter"
  /support/tickets:
    post:
      tags: [Support]
      operationId: createSupportTicket
      summary: Create a support ticket/contact request
      x-queue:
        name: notification-dispatch
        reason: Notify support/admins without blocking the request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportTicketCreateDto"
      responses:
        "201":
          description: Support ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
        "400":
          $ref: "#/components/responses/ValidationError"
  /feedback:
    post:
      tags: [Support]
      operationId: createFeedback
      summary: Submit product feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackCreateDto"
      responses:
        "201":
          description: Feedback captured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageQuery:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitQuery:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    SortOrderQuery:
      in: query
      name: sortOrder
      schema:
        type: string
        enum: [asc, desc]
        default: asc
...
    VehicleIdPath:
      in: path
      name: vehicleId
      required: true
      schema:
        type: string
        format: uuid
...
    PaymentIdPath:
      in: path
      name: paymentId
      required: true
      schema:
        type: string
...
    JobIdPath:
      in: path
      name: jobId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Conflict:
      description: Request conflicts with current state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    ApiError:
      type: object
      required: [statusCode, code, message]
      properties:
        statusCode:
          type: integer
          example: 400
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Please enter a valid email address.
        details:
          type: array
          items:
            $ref: "#/components/schemas/ApiErrorDetail"
        requestId:
          type: string
          nullable: true
    ApiErrorDetail:
      type: object
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: Please enter a valid email address.
...
    Pagination:
      type: object
      required: [total, page, limit, pages]
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer
    BatchMutationResult:
      type: object
      required: [requested, succeeded, failed]
      properties:
        requested:
          type: integer
        succeeded:
          type: integer
        failed:
          type: integer
        failedIds:
          type: array
          items:
            type: string
...
    Address:
      type: object
      required: [country, streetAddress, city, state, zipCode]
      properties:
        country:
          type: string
          example: Saudi Arabia
        streetAddress:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        lat:
          type: number
          format: double
          nullable: true
        lng:
          type: number
          format: double
          nullable: true
...
    MfaMethod:
      type: string
      enum: [sms, authenticator]
...
    MfaUpdateDto:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
        method:
          $ref: "#/components/schemas/MfaMethod"
        phone:
          type: string
          nullable: true
          pattern: "^\\+?[0-9\\s()-]{8,20}$"
...
    OnboardingStepId:
      type: string
      enum:
        - getting-started
        - owner-information
        - business-information
        - bank-details
        - tax-information
        - two-factor-auth
        - review
        - submitted
    OnboardingProgress:
      type: object
      required: [currentStep, completedSteps, percent, submitted]
      properties:
        currentStep:
          $ref: "#/components/schemas/OnboardingStepId"
        completedSteps:
          type: array
          items:
            $ref: "#/components/schemas/OnboardingStepId"
        percent:
          type: integer
          minimum: 0
          maximum: 100
        submitted:
          type: boolean
        missingRequirements:
          type: array
          items:
            type: string
    OnboardingStepResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        progress:
          $ref: "#/components/schemas/OnboardingProgress"
        nextStep:
          $ref: "#/components/schemas/OnboardingStepId"
    OwnerInformationDto:
      type: object
      required: [firstName, lastName, email, location]
      properties:
        firstName:
          type: string
          pattern: "^[A-Za-z]+$"
        lastName:
          type: string
          pattern: "^[A-Za-z]+$"
        email:
          type: string
          format: email
        location:
          $ref: "#/components/schemas/Address"
    BusinessType:
      type: string
      enum: [general-repair, specialized, body-paint, quick-service]
    EmployeeCount:
      type: string
      enum: [small, medium, large]
    BusinessInformationDto:
      type: object
      required: [businessName, businessType, employeeCount, location]
      properties:
        businessName:
          type: string
        businessType:
          $ref: "#/components/schemas/BusinessType"
        employeeCount:
          $ref: "#/components/schemas/EmployeeCount"
        location:
          $ref: "#/components/schemas/Address"
...
    Vehicle:
      type: object
      required: [id, customerId, make, model, year]
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        make:
          type: string
        model:
          type: string
        year:
          type: string
        plate:
          type: string
          nullable: true
        vin:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    VehicleCreateDto:
      type: object
      required: [customerId, make, model, year]
      properties:
        customerId:
          type: string
          format: uuid
        make:
          type: string
        model:
          type: string
        year:
          type: string
        plate:
          type: string
        vin:
          type: string
    VehicleUpdateDto:
      type: object
      properties:
        customerId:
          type: string
          format: uuid
        make:
          type: string
        model:
          type: string
        year:
          type: string
        plate:
          type: string
        vin:
          type: string
    PaginatedVehicles:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Vehicle"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    PaymentMethod:
      type: string
      enum: [VISA, Fawry, PayPal, Cash]
    PaymentKind:
      type: string
      enum: [item, service, car]
    PaymentStatus:
      type: string
      enum: [paid, pending, failed, refunded, partially_refunded]
    Payment:
      type: object
      required: [id, customer, item, method, date, amount, kind, status]
      properties:
        id:
          type: string
          example: PAY-8821
        businessId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
          nullable: true
        customer:
          type: string
        bookingId:
          type: string
          nullable: true
        inventoryItemId:
          type: string
          format: uuid
          nullable: true
        item:
          type: string
          description: Display name for the item, service, or car payment.
        method:
          $ref: "#/components/schemas/PaymentMethod"
        date:
          type: string
          format: date
        amount:
          type: number
          format: double
        currency:
          type: string
          default: USD
        kind:
          $ref: "#/components/schemas/PaymentKind"
        status:
          $ref: "#/components/schemas/PaymentStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PaymentSummary:
      type: object
      required: [monthlyPayments, yearlyPayments, itemsSold, servicesDone]
      properties:
        monthlyPayments:
          type: number
          format: double
        yearlyPayments:
          type: number
          format: double
        itemsSold:
          type: integer
        servicesDone:
          type: integer
        currency:
          type: string
          default: USD
    PaginatedPayments:
      type: object
      required: [data, summary, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        summary:
          $ref: "#/components/schemas/PaymentSummary"
        pagination:
          $ref: "#/components/schemas/Pagination"
    RefundPaymentDto:
      type: object
      properties:
        amount:
          type: number
          format: double
          nullable: true
          description: Omit for full refund.
        reason:
          type: string
          nullable: true
    BatchRefundPaymentsDto:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          minItems: 1
          items:
            type: string
        reason:
          type: string
          nullable: true
    PaymentRefund:
      type: object
      required: [id, paymentId, amount, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        paymentId:
          type: string
        amount:
          type: number
          format: double
        reason:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, succeeded, failed]
        createdAt:
          type: string
          format: date-time
    PaymentExportFormat:
      type: string
      enum: [excel, pdf, word, google-docs, dashboard]
    PaymentExportCreateDto:
      type: object
      required: [format, categories]
      properties:
        format:
          $ref: "#/components/schemas/PaymentExportFormat"
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        categories:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/PaymentKind"
        minPrice:
          type: number
          format: double
          nullable: true
        maxPrice:
          type: number
          format: double
          nullable: true
        paymentIds:
          type: array
          items:
            type: string
    ExportJobStatus:
      type: string
      enum: [queued, processing, completed, failed, expired]
    ExportJob:
      type: object
      required: [id, type, status, format, createdAt]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [payment-export]
        status:
          $ref: "#/components/schemas/ExportJobStatus"
        format:
          $ref: "#/components/schemas/PaymentExportFormat"
        filters:
          $ref: "#/components/schemas/PaymentExportCreateDto"
        matchedCount:
          type: integer
          nullable: true
        downloadUrl:
          type: string
          nullable: true
        errorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
...
    BankAccountCreateDto:
      type: object
      required: [bankName, accountHolder, accountNumber, iban, swift, branch, billingEmail]
      properties:
        bankName:
          type: string
        accountHolder:
          type: string
          pattern: "^[A-Za-z\\s]+$"
        accountNumber:
          type: string
          minLength: 6
          maxLength: 34
          pattern: "^[0-9]+$"
        iban:
          type: string
          pattern: "^[A-Z]{2}\\d{2}[A-Z0-9]{11,30}$"
        swift:
          type: string
          pattern: "^[A-Z]{4}[A-Z]{2}[A-Z0-9]{2}([A-Z0-9]{3})?$"
        branch:
          type: string
          pattern: "^[A-Za-z0-9\\s.'&-]+$"
        billingEmail:
          type: string
          format: email
...
    HelpCenter:
      type: object
      required: [categories]
      properties:
        categories:
          type: array
          items:
            $ref: "#/components/schemas/HelpCenterCategory"
    HelpCenterCategory:
      type: object
      required: [id, title, articles]
      properties:
        id:
          type: string
        title:
          type: string
        articles:
          type: array
          items:
            $ref: "#/components/schemas/HelpCenterArticle"
    HelpCenterArticle:
      type: object
      required: [id, title, body]
      properties:
        id:
          type: string
        title:
          type: string
        body:
          type: string
    SupportTicketCreateDto:
      type: object
      required: [subject, message]
      properties:
        subject:
          type: string
        message:
          type: string
        email:
          type: string
          format: email
          nullable: true
        priority:
          type: string
          enum: [low, normal, high]
          default: normal
    SupportTicket:
      type: object
      required: [id, subject, message, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        subject:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [open, pending, resolved, closed]
        createdAt:
          type: string
          format: date-time
    FeedbackCreateDto:
      type: object
      required: [message]
      properties:
        message:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        page:
          type: string
          nullable: true
    Feedback:
      type: object
      required: [id, message, createdAt]
      properties:
        id:
          type: string
          format: uuid
        message:
          type: string
        rating:
          type: integer
          nullable: true
        page:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...